Extract `pkg/sharding/key` package
